llm server/pyproject.toml
[tool.poetry]
name = "llm server to handle query request"
version = "0.0.1"
description = "LLM Server to Handle Query Request, supports all kind of LLM Managed or Unmanaged. Giving user a common interface to interact with Any LLM (OpenAI, Azure OpenAI, Vertex AI etc)"
authors = ["Harsh Raj <harsh.raj@scikiq.com>"]
license = "copyright@scikiq.com"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
openai = "^1.78.1"
requests = "^2.32.3"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
